+++ src/containers/card/CardPasswordInputContainer.tsx
@@ -1,9 +1,9 @@
 import React, { useRef } from 'react';
-import { useAppDispatch, useAppState } from '../../../hooks/hooks';
-import { ActionType } from '../../../types';
-import { createAction } from '../../../context/Provider';
-import CardFormInput from '../CardFormInput';
 import { css } from '@emotion/react';
+import { useAppDispatch, useAppState } from 'hooks/hooks';
+import { createAction } from 'context/Provider';
+import { ActionType } from 'types';
+import CardFormInput from 'components/card/CardFormInput';
 
 const style = css({
   height: '45px',
